+++ src/mpd/mpd_client.rs
@@ -61,6 +61,7 @@ impl ValueChange {
 #[allow(dead_code)]
 pub trait MpdClient {
     fn version(&mut self) -> Version;
+    fn binary_limit(&mut self, limit: u64) -> MpdResult<()>;
     fn password(&mut self, password: &str) -> MpdResult<()>;
     fn commands(&mut self) -> MpdResult<MpdList>;
     fn update(&mut self, path: Option<&str>) -> MpdResult<Update>;
@@ -135,6 +136,11 @@ pub trait MpdClient {
 impl MpdClient for Client<'_> {
     fn version(&mut self) -> Version {
         self.version
+    }
+
+    fn binary_limit(&mut self, limit: u64) -> MpdResult<()> {
+        self.send(&format!("binarylimit {limit}"))
+            .and_then(ProtoClient::read_ok)
     }
 
     fn password(&mut self, password: &str) -> MpdResult<()> {